local_setup/requirements.txt
# Local setup dependencies
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy>=1.24.0
